+++ example/yafsm_example.dart
@@ -1,20 +1,19 @@
 import 'package:yafsm/yafsm.dart';
 
-class SwitchMachine {
-  SwitchMachine() {
-    root.initialize(isOff);
+class SwitchMachine extends Machine {
+  SwitchMachine() : super('Switch') {
+    initialize(isOff);
   }
 
-  final root = Machine('switch');
-  late final isOn = root.state('on');
-  late final isOff = root.state('off');
-  late final turnOn = root.transition('turn on', {isOff}, isOn);
-  late final turnOff = root.transition('turn off', {isOn}, isOff);
+  late final isOn = state('on');
+  late final isOff = state('off');
+  late final turnOn = transition('turn on', {isOff}, isOn);
+  late final turnOff = transition('turn off', {isOn}, isOff);
 }
 
 void main() {
   final m = SwitchMachine();
-  m.root.start();
+  m.start();
   print(m.isOn()); // -> false
   m.turnOn();
   print(m.isOn()); // -> true